added category popup

+++ app/src/main/java/com/redc4ke/taniechlanie/data/DataOperations.kt
@@ -69,7 +69,7 @@ fun priceString(price: BigDecimal, context: Context): String {
     )
 }
 
-fun volumeString(volume: Long, context: Context): String {
+fun volumeString(volume: Int, context: Context): String {
     return context.getString(
         R.string.suff_volume,
         volume.toString()
